+++ changelog.md
@@ -16,6 +16,13 @@ All notable changes to this project  be documented in this file.
 
 - nothing so far 
 -->
+
+## [Unreleased] - 2022-Nov-25
+
+### Changed
+
+- Change the formatting of context manager checks such that if a msg is included, we see both the message and the assert failure. Previoulsy, the message would replace the assert failure.
+
 ## [1.2.0] - 2022-Nov-25
 
 ### Added